requirements.txt
--find-links wheelhouse
arrow==0.8.0
azure-common==1.1.4
azure-nspkg==1.0.0
azure-servicebus==0.20.2
Babel==2.3.4
blinker==1.4
click==6.6
Flask==0.11.1
Flask-Babel==0.11.1
Flask-CDN==1.5.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Session==0.2.3
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
ptvsd==2.2.0
pymssql==2.1.3
PyMySQL==0.7.5
python-dateutil==2.5.3
pytz==2016.4
redis==2.10.5
requests==2.10.0
six==1.10.0
SQLAlchemy==1.0.14
times==0.7
Werkzeug==0.11.10
WTForms==2.1
